Unify multi remote test and moved them to functional
